poly1305: remove `Mac` impl and refactor
As a universal hash function and one-time authenticator, Poly1305
doesn't meet normal expectations of a MAC.

This change removes the `Mac` impl, but otherwise leaves the API looking
very much like a `Mac`.

Additionally, this commit includes a lot of refactoring in an attempt to
clean up the code somewhat.