poly1305: Add `input_padded` method
Add a function for inputting data into Poly1305 which is padded to its
block size, a.k.a. the `pad16()` function from RFC 8439.

This is useful for implementing authenticated encryption constructions
using Salsa20 family ciphers (i.e. Salsa20Poly1305, ChaCha20Poly1305)
which is the main use case for Poly1305 to begin with.