.travis.yml: Bump `polyval` MSRV to 1.32.0
The POLYVAL tests are failing on Rust 1.27.0, but passing on newer
versions, suggesting there may be something amiss with `core::arch`
support for `pclmulqdq` on older Rust versions.

This bumps it up to our ideal MSRV, as 1.32.0+ would allow us to replace
the dependency on the `byteorder` crate with equivalent `core`
functions.