hmac: bump `digest` to v0.11.0-pre.3; MSRV 1.71
This prerelease of `digest` notably includes a migration from
`generic-array` to `hybrid-array`.

Bumps the crate version to v0.13.0-pre to denote this is a breaking
change, however note there is not a crate release associated with this
version bump.

The edition has also been bumped to 2021, and `hex-literal` has been
upgraded to v0.4.